requirements.txt
amqp==2.2.2
astroid==1.5.3
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
colorful==0.4.0
defusedxml==0.5.0
Django==1.11.6
django-allauth==0.34.0
django-appconf==1.0.2
django-bootstrap4==0.0.4
django-celery-email==2.0.0
docopt==0.6.2
humanize==0.5.1
idna==2.6
isort==4.2.15
kombu==4.1.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.12
oauthlib==2.0.6
parse==1.8.2
pep8==1.7.1
PyHamcrest==1.9.0
pylint==1.7.4
pysingleton==0.2.1
python3-openid==3.1.0
pytz==2017.3
PyYAML==3.12
radish-bdd==0.8.0
radish-parse-type==0.3.5
requests==2.18.4
requests-oauthlib==0.8.0
selenium==3.7.0
six==1.11.0
tag-expressions==1.1.0
urllib3==1.22
vine==1.1.4
wrapt==1.10.11
